Fixed: the issue of empty cards on the orders page and removed the usage of order card component(#1ym1mnp)
